requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.0
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Deprecated==1.2.14
fastapi==0.115.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
idna==3.10
imageio==2.35.1
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.12
requests==2.32.3
rich==13.9.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.5.0
tf-keras-vis==0.8.7
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
